my-ai-workspace/ai-agent/pyproject.toml
[tool.poetry]
name = "ai-agent"
version = "0.1.0"
description = "FastAPI service generated with RapidKit"
authors = ["rapidx <you@example.com>"]
license = "MIT"
packages = [{ include = "src" }]

[tool.poetry.scripts]
dev = "src.cli:dev"
start = "src.cli:start"
build = "src.cli:build"
test = "src.cli:test"
lint = "src.cli:lint"
format = "src.cli:format"
help = "src.cli:help_cmd"

[tool.poetry.dependencies]
python             = "^3.10"
fastapi            = "^0.128.0"
uvicorn            = { extras = ["standard"], version = "^0.40.0" }
pydantic           = "^2.12.5"
pydantic-settings  = { version = "^2.12.0", extras = ["yaml"] }
# <<<inject:poetry-dependencies>>>
# <<<inject:module-dependencies>>>
boto3              = ">=1.35.0,<2.0"
docker             = "7.1.0"
httpx              = ">=0.28.0,<1.0"
hvac               = ">=2.3.0,<3.0"
python-json-logger = ">=2.0.0,<3.0"
pyyaml             = ">=6.0.2,<7.0"
yamllint           = ">=1.35.0,<2.0"
openai = "^2.20.0"
[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"
httpx = "^0.28.1"
ruff = "^0.14.10"
black = "^25.12.0"
isort = "^7.0.0"
mypy = "^1.19.1"
build = "^1.3.0"

# <<<inject:poetry-dev-dependencies>>>

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
	"integration: mark tests as integration tests",
	"template_integration: mark template integration tests",
]

[tool.ruff]
ignore = ["E501"]
line-length = 100
exclude = [".rapidkit/vendor"]

[tool.mypy]
python_version = "3.10.14"
ignore_missing_imports = true
warn_unused_configs = false
show_error_codes = true
# Ignore vendor snapshots and generated health shims from strict static checking
exclude = "(^\\.rapidkit/vendor/|src/health/|src/main.py$)"